% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runners.R
\name{annotate_demo_result}
\alias{annotate_demo_result}
\title{Mark a result as a demo result}
\usage{
annotate_demo_result(x)
}
\arguments{
\item{x}{A result object.}
}
\value{
\code{x} with a \code{llmrshiny_demo} attribute.
}
\description{
Mark a result as a demo result
}
