% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{column_names_for_mapping}
\alias{column_names_for_mapping}
\title{Column names of a data frame, for a mapping select input}
\usage{
column_names_for_mapping(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A character vector of column names.
}
\description{
Column names of a data frame, for a mapping select input
}
