% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{cost_tile}
\alias{cost_tile}
\title{A value box reporting session usage}
\usage{
cost_tile(state)
}
\arguments{
\item{state}{A cost record.}
}
\value{
A \code{bslib::value_box}.
}
\description{
A value box reporting session usage
}
