% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guards.R
\name{github_remote_for}
\alias{github_remote_for}
\title{GitHub remote for an LLMR-family package}
\usage{
github_remote_for(package)
}
\arguments{
\item{package}{Package name.}
}
\value{
A \code{"owner/repo"} string for \code{remotes::install_github()}.
}
\description{
GitHub remote for an LLMR-family package
}
