% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runners.R
\name{is_demo_result}
\alias{is_demo_result}
\title{Is a result a demo result?}
\usage{
is_demo_result(x)
}
\arguments{
\item{x}{A result object.}
}
\value{
\code{TRUE} when marked by \code{\link[=annotate_demo_result]{annotate_demo_result()}}.
}
\description{
Is a result a demo result?
}
