% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guards.R
\name{pkg_available}
\alias{pkg_available}
\title{Is a package installed?}
\usage{
pkg_available(package)
}
\arguments{
\item{package}{Package name.}
}
\value{
\code{TRUE} if the package can be loaded.
}
\description{
Is a package installed?
}
