% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{provider_choices}
\alias{provider_choices}
\title{Provider choices for a select input}
\usage{
provider_choices()
}
\value{
A named character vector (\code{display} -> \code{provider}).
}
\description{
Provider choices for a select input
}
