% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{provider_display_name}
\alias{provider_display_name}
\title{Display name for a provider}
\usage{
provider_display_name(provider)
}
\arguments{
\item{provider}{Provider id.}
}
\value{
The display name, or the provider id.
}
\description{
Display name for a provider
}
