% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage.R
\name{usage_tile}
\alias{usage_tile}
\title{A value box reporting session usage}
\usage{
usage_tile(state)
}
\arguments{
\item{state}{A usage record.}
}
\value{
A \code{bslib::value_box}.
}
\description{
A value box reporting session usage
}
