% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{validate_column_mapping}
\alias{validate_column_mapping}
\title{Validate a column mapping}
\usage{
validate_column_mapping(data, text_col, label_col = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{text_col}{Required text column name.}

\item{label_col}{Optional label column name.}
}
\value{
\code{TRUE}, or an error.
}
\description{
Validate a column mapping
}
