% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runners.R
\name{is_demo_result}
\alias{is_demo_result}
\title{Is a result a demonstration result?}
\usage{
is_demo_result(x)
}
\arguments{
\item{x}{A result object.}
}
\value{
\code{TRUE} when the result has the demonstration class or source fields.
}
\description{
Is a result a demonstration result?
}
