% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLMR.shiny-package.R
\docType{package}
\name{LLMR.shiny-package}
\alias{LLMR.shiny}
\alias{LLMR.shiny-package}
\title{LLMR.shiny: Shared 'Shiny' Substrate for 'LLMR' Family GUIs}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The shell from which the 'LLMR' family's graphical interfaces are assembled. Interface conventions are implemented once here, so a fix reaches every interface built on this package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asanaei/LLMR.shiny}
  \item Report bugs at \url{https://github.com/asanaei/LLMR.shiny/issues}
}

}
\author{
\strong{Maintainer}: Ali Sanaei \email{sanaei@uchicago.edu}

}
\keyword{internal}
