% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLMR.shiny-package.R
\docType{package}
\name{LLMR.shiny-package}
\alias{LLMR.shiny}
\alias{LLMR.shiny-package}
\title{LLMR.shiny: Shared 'Shiny' Components for 'LLMR' Family Applications}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Reusable 'Shiny' user interface and server components from which the graphical applications in the 'LLMR' package family are assembled.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asanaei/LLMR.shiny}
  \item Report bugs at \url{https://github.com/asanaei/LLMR.shiny/issues}
}

}
\author{
\strong{Maintainer}: Ali Sanaei \email{sanaei@uchicago.edu}

}
\keyword{internal}
